[Chore] 사용자 계정별 알림 진행중
