+++ src/lib/actions/userAction.ts
@@ -1,5 +1,5 @@
 import connectDB from "../db";
-import { Alert, User } from "../schema";
+import { User } from "../schema";
 
 export async function getUserData(userId: string) {
   await connectDB();
@@ -15,28 +15,5 @@ export async function getUserData(userId: string) {
   } catch (error: any) {
     console.error("error", error.message);
     return { state: false, message: "잘못된 시도입니다. userId: " + userId };
-  }
-}
-
-export async function updateAlert(userId: string, formData: FormData) {
-  const postComment = formData.get("postComment");
-  const studyComment = formData.get("studyComment");
-
-  await connectDB();
-
-  try {
-    const data = await Alert.findOneAndUpdate(
-      { userId },
-      {
-        count: {
-          postComment,
-          studyComment,
-        },
-      }
-    );
-
-    return { state: true, data };
-  } catch (error) {
-    return { state: false, message: "알림 정보에 실패했습니다." };
   }
 }
